+++ ISort_Strategy.h
@@ -2,5 +2,5 @@
 #include <vector>
 class ISort_Strategy {
 public:
-	virtual void sort(int A, int n, const std::vector<int>& a, const std::vector<int>& C, std::vector<int>& result_indexes) = 0;
+	virtual void sort(short int A, short int n, const std::vector<short int>& a, const std::vector<short int>& C, std::vector<short int>& result_indexes) = 0;
 };
